pyproject.toml
[tool.poetry]
name = "ecommerce-scraping"
version = "0.1.0"
description = "Application to scrape products data from an e-commerce website."
authors = ["Federico Trifoglio <federico.trifoglio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
ipykernel = "^6.25.2"
aiohttp = "^3.8.5"
pandas = "^2.1.0"
requests = "^2.31.0"
loguru = "^0.7.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 88

[tool.ruff]
select = ["E", "F", "D", "I"]
ignore = ["D417"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["D"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.mypy]
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "loguru.*",
    "dotenv.*",
    "aiohttp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "./tests"
]
